requirements.txt
streamlit
pandas
openpyxl
fpdf2
plotly
Pillow
st-gsheets-connection
gspread